+++ contracts/interface/IClearingHouseConfig.sol
@@ -1,7 +1,25 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity 0.7.6;
 
-interface IClearingHouseConfig {
+interface IClearingHouseConfigEvent {
+    event LiquidationPenaltyRatioChanged(uint24 liquidationPenaltyRatio);
+
+    event PartialCloseRatioChanged(uint24 partialCloseRatio);
+
+    event TwapIntervalChanged(uint256 twapInterval);
+
+    event MaxMarketsPerAccountChanged(uint8 maxMarketsPerAccount);
+
+    event SettlementTokenBalanceCapChanged(uint256 cap);
+
+    event MaxFundingRateChanged(uint24 rate);
+
+    event MarkPriceMarketTwapIntervalChanged(uint32 twapInterval);
+
+    event MarkPricePremiumIntervalChanged(uint32 premiumInterval);
+}
+
+interface IClearingHouseConfig is IClearingHouseConfigEvent {
     /// @return maxMarketsPerAccount Max value of total markets per account
     function getMaxMarketsPerAccount() external view returns (uint8 maxMarketsPerAccount);
 
@@ -25,9 +43,6 @@ interface IClearingHouseConfig {
 
     /// @return maxFundingRate Max value of funding rate
     function getMaxFundingRate() external view returns (uint24 maxFundingRate);
-
-    /// @return isBackstopLiquidityProvider is backstop liquidity provider
-    function isBackstopLiquidityProvider(address account) external view returns (bool isBackstopLiquidityProvider);
 
     /// @return marketTwapInterval MarketTwapInterval is the interval of market twap used for mark price calculations
     /// @return premiumInterval PremiumInterval is the interval of premium used for mark price calculations